feat: add areas-forms POST-GET route

+++ src/modules/area/area.service.ts
@@ -11,18 +11,34 @@ export class AreaService {
     private prisma: PrismaService,
   ) { }
 
-  private async findAreaByName(name: string): Promise<areas> {
-    return await this.prisma.areas.findFirst({
+  async ExistsAreaByName(name: string): Promise<boolean> {
+    const areaFound = await this.prisma.areas.findFirst({
       where: { name: name }
     });
+    if (areaFound) return true;
+    return false;
+  }
+
+
+  async findAreaByName(name: string): Promise<areas> {
+    const areaFound = await this.prisma.areas.findFirst({
+      where: { name: name }
+    });
+    if (!areaFound) throw new BadRequestException('Area not found');
+    return areaFound;
   }
 
   async create(createAreaDto: CreateAreaDto): Promise<areas> {
-    const area = await this.findAreaByName(createAreaDto.name);
+    const area = await this.ExistsAreaByName(createAreaDto.name);
     if (area) throw new BadRequestException('Area already exists');
-    return await this.prisma.areas.create({
-      data: { name: createAreaDto.name }
-    });
+    try {
+      return await this.prisma.areas.create({
+        data: { name: createAreaDto.name }
+      });
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
+
   }
 
   async findAll(): Promise<areas[]> {
@@ -30,25 +46,40 @@ export class AreaService {
   }
 
   async findOne(id: string): Promise<areas> {
-    return await this.prisma.areas.findUnique({
-      where: { id: id }
-    })
+    const areaFound = await this.findOne(id);
+    if (!areaFound) throw new BadRequestException('Area not found');
+    try {
+      return await this.prisma.areas.findUnique({
+        where: { id: id }
+      })
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
   }
 
   async update(id: string, updateAreaDto: UpdateAreaDto) {
     const areaFound = await this.findOne(id);
     if (!areaFound) throw new BadRequestException('Area not found');
-    console.log("🚀 ~ AreaService ~ update ~ areaFound:", areaFound)
-    return await this.prisma.areas.update({
-      where: { id: areaFound.id },
-      data: { ...updateAreaDto }
-    });
+    try {
+      return await this.prisma.areas.update({
+        where: { id: areaFound.id },
+        data: { name: updateAreaDto.name }
+      });
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
   }
 
   async remove(id: string): Promise<areas> {
-    return await this.prisma.areas.delete({
-      where: { id: id }
-    });
+    const areaFound = await this.findOne(id);
+    if (!areaFound) throw new BadRequestException('Area not found');
+    try {
+      return await this.prisma.areas.delete({
+        where: { id: id }
+      });
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
   }
 
 }
